tests/wpt/tests/css/filter-effects/parsing/color-interpolation-filters-parsing-invalid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>Filter Effects Module Level 1: parsing color-interpolation-filters with invalid values</title>
<link rel="author" title="Eric Willigers" href="mailto:ericwilligers@chromium.org">
<link rel="help" href="https://drafts.fxtf.org/filter-effects/#ColorInterpolationFiltersProperty">
<meta name="assert" content="color-interpolation-filters supports only the grammar 'auto | sRGB | linearRGB'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_invalid_value("color-interpolation-filters", "none");
test_invalid_value("color-interpolation-filters", "linearRGB sRGB");
test_invalid_value("color-interpolation-filters", "auto sRGB linearRGB");
</script>
</body>
</html>
